+++ src/Indexing/Visiting/FunctionIndexingVisitor.php
@@ -186,7 +186,17 @@ final class FunctionIndexingVisitor extends NodeVisitorAbstract
                 $typeStringSpecification = $nodeType->name;
             }
 
-            $typeStringSpecification = $nodeType->toString();
+            $nodeTypes = [];
+
+            if ($nodeType instanceof Node\UnionType) {
+                $nodeTypes = array_map(function ($nodeType): string {
+                    return (string) $nodeType;
+                }, $nodeType->types);
+            } else {
+                $nodeTypes = [$nodeType->toString()];
+            }
+
+            $typeStringSpecification = implode('|', $nodeTypes);
 
             if ($node->getReturnType() instanceof Node\NullableType) {
                 $typeStringSpecification .= '|null';
@@ -288,6 +298,18 @@ final class FunctionIndexingVisitor extends NodeVisitorAbstract
                     $unresolvedType = new IdentifierTypeNode(NodeHelpers::fetchClassName($typeNode));
                 } elseif ($typeNode instanceof Node\Identifier) {
                     $unresolvedType = new IdentifierTypeNode($typeNode->name);
+                } else /*if ($typeNode instanceof Node\UnionType)*/ {
+                    $adaptedTypes = [];
+
+                    foreach ($typeNode->types as $nestedTypeNode) {
+                        if ($nestedTypeNode instanceof Node\Name) {
+                            $adaptedTypes[] = new IdentifierTypeNode(NodeHelpers::fetchClassName($nestedTypeNode));
+                        } else /*if ($nestedTypeNode instanceof Node\Identifier)*/ {
+                            $adaptedTypes[] = new IdentifierTypeNode($nestedTypeNode->name);
+                        }
+                    }
+
+                    $unresolvedType = new UnionTypeNode($adaptedTypes);
                 }
 
                 if ($param->type instanceof Node\NullableType) {
@@ -296,7 +318,7 @@ final class FunctionIndexingVisitor extends NodeVisitorAbstract
                         new IdentifierTypeNode(SpecialDocblockTypeIdentifierLiteral::NULL_),
                     ]);
                 } elseif ($param->default instanceof Node\Expr\ConstFetch &&
-                    $param->default->name->toString() === 'null'
+                    $param->default->name->toString() === SpecialDocblockTypeIdentifierLiteral::NULL_
                 ) {
                     $unresolvedType = new UnionTypeNode([
                         $unresolvedType,